updated sleep

+++ .github/scripts/validate-deploy.sh
@@ -49,6 +49,49 @@ validate_gitops_content "${NAMESPACE}" "${LAYER}" "${SERVER_NAME}" "${TYPE}" "${
 check_k8s_namespace "${NAMESPACE}"
 
 #check_k8s_resource "${NAMESPACE}" "deployment" "${COMPONENT_NAME}"
+if [[ ! -f "argocd/${LAYER}/cluster/${SERVER_NAME}/${TYPE}/${NAMESPACE}-${COMPONENT_NAME}.yaml" ]]; then
+  echo "ArgoCD config missing - argocd/${LAYER}/cluster/${SERVER_NAME}/${TYPE}/${NAMESPACE}-${COMPONENT_NAME}.yaml"
+  exit 1
+fi
+
+echo "Printing argocd/${LAYER}/cluster/${SERVER_NAME}/${TYPE}/${NAMESPACE}-${COMPONENT_NAME}.yaml"
+cat "argocd/${LAYER}/cluster/${SERVER_NAME}/${TYPE}/${NAMESPACE}-${COMPONENT_NAME}.yaml"
+
+if [[ ! -f "payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/values.yaml" ]]; then
+  echo "Application values not found - payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/values.yaml"
+  exit 1
+fi
+
+echo "Printing payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/values.yaml"
+cat "payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/values.yaml"
+
+count=0
+until kubectl get namespace "${NAMESPACE}" 1> /dev/null 2> /dev/null || [[ $count -eq 20 ]]; do
+  echo "Waiting for namespace: ${NAMESPACE}"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for namespace: ${NAMESPACE}"
+  exit 1
+else
+  echo "Found namespace: ${NAMESPACE}. Sleeping for 30 seconds to wait for everything to settle down"
+  sleep 30
+fi
+
+## Check if the icp4a-root-ca is there 
+count=0
+until kubectl get secret icp4a-root-ca -n "${NAMESPACE}" || [[ $count -eq 60 ]]; do
+  echo "Waiting for secret icp4a-root-ca in ${NAMESPACE} COUNTER $count" 
+  count=$((count + 1))
+  sleep 60
+done
+
+if [[ $count -eq 60 ]]; then
+  echo "Timed out waiting for icp4a-root-ca in ${NAMESPACE}"
+  exit 1
+fi
 
 ## Check if the admin.registrykey is there 
 count=0
@@ -57,6 +100,11 @@ until kubectl get secret admin.registrykey -n "${NAMESPACE}" || [[ $count -eq 20
   count=$((count + 1))
   sleep 15
 done
+
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for admin.registrykey in ${NAMESPACE}"
+  exit 1
+fi
 
 ## Check if the subscription for ibm-automation is there 
 SUBSNAME="ibm-automation"
@@ -68,20 +116,28 @@ until kubectl get subs -n "${NAMESPACE}" |grep "${SUBSNAME}" || [[ $count -eq 20
   sleep 15
 done
 
-## Check if the icp4a-root-ca is there 
+if [[ $count -eq 20 ]]; then
+  echo "Timed out waiting for Subscription/${SUBSNAME} in ${NAMESPACE}"
+  exit 1
+fi
+
+## Check if the configmaps is there 
 count=0
-until kubectl get secret icp4a-root-ca -n "${NAMESPACE}" || [[ $count -eq 30 ]]; do
-  echo "Waiting for secret icp4a-root-ca in ${NAMESPACE} COUNTER $count" 
+until kubectl get configmaps icp4adeploy-gitops-cp-odm-access-info -n "${NAMESPACE}" || [[ $count -eq 30 ]]; do
+  echo "Waiting for configmaps icp4adeploy-gitops-cp-odm-access-info in ${NAMESPACE} COUNTER $count" 
   count=$((count + 1))
   sleep 40
 done
 
+if [[ $count -eq 30 ]]; then
+  echo "Timed out waiting for icp4adeploy-gitops-cp-odm-access-info in ${NAMESPACE}"
+  exit 1
+fi
 
 #### Temporary sleep to validate deployment manually
 count=0
 echo "Sleeping for 10 minutes after finding the subscription to manually verify"
-sleep 1200
-#
+sleep 600
 
 cd ..
 rm -rf .testrepo
